max_align_t: apply alignof to 'long double' for long double alignment
Reviewed by;	markj
Sponsored by:	The FreeBSD Foundation
MFC after:	1 week
Differential revision:	https://reviews.freebsd.org/D53826

+++ sys/sys/_types.h
@@ -183,7 +183,7 @@ typedef struct {
 	    __attribute__((__aligned__(__alignof__(long long))));
 #ifndef _STANDALONE
 	long double __max_align2
-	    __attribute__((__aligned__(__alignof__(long long))));
+	    __attribute__((__aligned__(__alignof__(long double))));
 #endif
 } __max_align_t;
